Clarify ambiguity in encryption spec/example
From #43, as well as [packets.go](1), the final flag should be the first
item in each payload packet in an encrypted message. However, the
example currently places the final flag last in the payload packet,
introducing ambiguity for those implementing the protocol. This commit
changes the example to confirm to the actual implementation behaviour.

1: https://github.com/keybase/saltpack/blob/e19b1910c0c5e2e309b248ea32d1d05e6b868dc4/packets.go#L62